+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Below are the most important changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ## 0.8.3 (2021-Sep-??)
-  - ...
+### Noteworthy code-changes
+  - This release drops Python 3.6 support (issue #3459)
 
 
 ## 0.8.2 (2021-Sep-01)